Fix compile errors from darside-tests module
