--- darkside-tests/tests/advanced_reorg_tests_constants.rs
@@ -1,8 +1,0 @@
-pub const ADVANCED_REORG_TESTS_USER_WALLET: &'static str =
-    "wrong when collect sponsor always simple control color exercise dad merry diet script attract public lucky pen pistol depend deposit salad room similar hour";
-pub const REORG_CHANGES_INCOMING_TX_HEIGHT_AFTER: &'static str =
-    "tests/data/wallets/advanced_reorg_tests/tx_height_changes/after_reorg.txt";
-pub const REORG_CHANGES_INCOMING_TX_HEIGHT_BEFORE: &'static str =
-    "tests/data/wallets/advanced_reorg_tests/tx_height_changes/before_reorg.txt";
-pub const TREE_STATE_FOLDER_PATH: &'static str =
-    "tests/data/wallets/advanced_reorg_tests/tx_height_changes/tree_states";
